+++ test/app_info_flutter_method_channel_test.dart
@@ -11,7 +11,18 @@ void main() {
   setUp(() {
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(channel, (MethodCall methodCall) async {
-          return '42';
+          switch (methodCall.method) {
+            case 'getAppName':
+              return 'AppInfo';
+            case 'getPackageName':
+              return 'com.example.app';
+            case 'getVersion':
+              return '1.0.0';
+            case 'getBuildNumber':
+              return '1';
+            default:
+              return null;
+          }
         });
   });
 
@@ -20,7 +31,19 @@ void main() {
         .setMockMethodCallHandler(channel, null);
   });
 
-  test('getPlatformVersion', () async {
-    expect(await platform.getPlatformVersion(), '42');
+  test('getAppName', () async {
+    expect(await platform.getAppName(), 'AppInfo');
+  });
+
+  test('getPackageName', () async {
+    expect(await platform.getPackageName(), 'com.example.app');
+  });
+
+  test('getVersion', () async {
+    expect(await platform.getVersion(), '1.0.0');
+  });
+
+  test('getBuildNumber', () async {
+    expect(await platform.getBuildNumber(), '1');
   });
 }