+++ test/app_info_flutter_test.dart
@@ -8,21 +8,55 @@ class MockAppInfoFlutterPlatform
     with MockPlatformInterfaceMixin
     implements AppInfoFlutterPlatform {
   @override
-  Future<String?> getPlatformVersion() => Future.value('42');
+  Future<String?> getAppName() => Future.value('AppInfo');
+
+  @override
+  Future<String?> getPackageName() => Future.value('com.example.app');
+
+  @override
+  Future<String?> getVersion() => Future.value('1.0.0');
+
+  @override
+  Future<String?> getBuildNumber() => Future.value('1');
 }
 
 void main() {
-  final AppInfoFlutterPlatform initialPlatform = AppInfoFlutterPlatform.instance;
+  final AppInfoFlutterPlatform initialPlatform =
+      AppInfoFlutterPlatform.instance;
 
   test('$MethodChannelAppInfoFlutter is the default instance', () {
     expect(initialPlatform, isInstanceOf<MethodChannelAppInfoFlutter>());
   });
 
-  test('getPlatformVersion', () async {
+  test('getAppName', () async {
     AppInfoFlutter appInfoFlutterPlugin = AppInfoFlutter();
     MockAppInfoFlutterPlatform fakePlatform = MockAppInfoFlutterPlatform();
     AppInfoFlutterPlatform.instance = fakePlatform;
 
-    expect(await appInfoFlutterPlugin.getPlatformVersion(), '42');
+    expect(await appInfoFlutterPlugin.getAppName(), 'AppInfo');
+  });
+
+  test('getPackageName', () async {
+    AppInfoFlutter appInfoFlutterPlugin = AppInfoFlutter();
+    MockAppInfoFlutterPlatform fakePlatform = MockAppInfoFlutterPlatform();
+    AppInfoFlutterPlatform.instance = fakePlatform;
+
+    expect(await appInfoFlutterPlugin.getPackageName(), 'com.example.app');
+  });
+
+  test('getVersion', () async {
+    AppInfoFlutter appInfoFlutterPlugin = AppInfoFlutter();
+    MockAppInfoFlutterPlatform fakePlatform = MockAppInfoFlutterPlatform();
+    AppInfoFlutterPlatform.instance = fakePlatform;
+
+    expect(await appInfoFlutterPlugin.getVersion(), '1.0.0');
+  });
+
+  test('getBuildNumber', () async {
+    AppInfoFlutter appInfoFlutterPlugin = AppInfoFlutter();
+    MockAppInfoFlutterPlatform fakePlatform = MockAppInfoFlutterPlatform();
+    AppInfoFlutterPlatform.instance = fakePlatform;
+
+    expect(await appInfoFlutterPlugin.getBuildNumber(), '1');
   });
 }